@@ -1,10 +1,10 @@
 apply plugin: 'com.android.application'
 
 android {
-    compileSdk 32
+    compileSdk 33
     defaultConfig {
         applicationId "com.topmas.top"
-        minSdkVersion 24
+        minSdkVersion 28
         targetSdkVersion 33
         versionCode 40   // TODO aqui se cambia la versión de la APP
         versionName "2"
