@@ -1,40 +1,29 @@
 package com.topmas.top;
 
-import static com.topmas.top.Constants.TAG_DESCRIPCIONINCIDENCIA;
 import static com.topmas.top.Constants.TAG_DIRECCION;
 import static com.topmas.top.Constants.TAG_ERROR;
-import static com.topmas.top.Constants.TAG_FECHAHORA;
-import static com.topmas.top.Constants.TAG_FECHAHORARESPUESTA;
+import static com.topmas.top.Constants.TAG_FUENTE;
 import static com.topmas.top.Constants.TAG_IDINC;
 import static com.topmas.top.Constants.TAG_IDPROMOTOR;
 import static com.topmas.top.Constants.TAG_IDRUTA;
 import static com.topmas.top.Constants.TAG_LATITUD;
 import static com.topmas.top.Constants.TAG_LONGITUD;
-import static com.topmas.top.Constants.TAG_OBSERVACIONES;
-import static com.topmas.top.Constants.TAG_RESPUESTA;
-import static com.topmas.top.Constants.TAG_SERVIDOR;
 import static com.topmas.top.Constants.TAG_TIENDA;
 import static com.topmas.top.Constants.TAG_USUARIO;
 
 import android.Manifest;
 import android.annotation.SuppressLint;
-import android.app.ProgressDialog;
 import android.content.Intent;
 import android.content.SharedPreferences;
 import android.content.pm.PackageManager;
-import android.database.Cursor;
-import android.location.Location;
 import android.os.Build;
 import android.os.Bundle;
 import android.preference.PreferenceManager;
 import android.util.Log;
 import android.view.View;
-import android.widget.AdapterView;
 import android.widget.Button;
 import android.widget.EditText;
 import android.widget.ImageView;
-import android.widget.SimpleCursorAdapter;
-import android.widget.Spinner;
 import android.widget.TextView;
 
 import androidx.appcompat.app.AppCompatActivity;
@@ -42,12 +31,8 @@ import androidx.core.app.ActivityCompat;
 
 import com.google.android.gms.location.FusedLocationProviderClient;
 import com.google.android.gms.location.LocationServices;
-import com.google.android.gms.tasks.OnSuccessListener;
 import com.google.android.material.floatingactionbutton.FloatingActionButton;
 import com.topmas.top.Objetos.oRespuestaIncidencia;
-
-import java.text.SimpleDateFormat;
-import java.util.Date;
 
 
 // FUENTE chat gpt busqueda de archivos de capturas de pantalla
@@ -60,27 +45,19 @@ public class RespuestaIncidencia extends AppCompatActivity {
     String pdireccion = "";
     String ptienda = "";
     String pUsuario = "";
-    String pfechahora = "";
-    String pfechahora_respuesta = "";
-    String prespuesta = "";
-    String pobservaciones = "";
+    String pfuente = "";
+
     AlmacenaImagen almacenaImagen;
-    private final Funciones funciones = new Funciones();
+
     private final Usuario usr = new Usuario();
-    private ImageView imageView;
+    ImageView imageView;
     Button btnPaste;
     private static final int REQUEST_CODE_PICK_IMAGE = 100;
-    ProgressDialog pDialog;
-    public static final String UPLOAD_INCIDENCIA = TAG_SERVIDOR + "/PhotoUpload/upload_incidencia.php";
-    public static final String UPLOAD_INCIDENCIA_O = TAG_SERVIDOR + "/PhotoUpload/upload_incidencia_o.php";
-    public static final String UPLOAD_IDINCIDENCIA = "idincidencia";
-    public static final String UPLOAD_OBSERVACIONES = "observaciones";
-    int iFoto = 0;            // Es el id de la tabla almacenfoto de la foto recien subida
-    int idoperacion = 11;
-    int idincidencia = 0;
-    private FusedLocationProviderClient fusedLocationClient;
 
-    @SuppressLint("MissingInflatedId")
+
+    FusedLocationProviderClient fusedLocationClient;
+
+    @SuppressLint({"MissingInflatedId", "ObsoleteSdkInt"})
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
@@ -91,47 +68,46 @@ public class RespuestaIncidencia extends AppCompatActivity {
         imageView = findViewById(R.id.imageClipboard);
         btnPaste = findViewById(R.id.btnPaste);
 
-
+/*
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.M) {
             if (checkSelfPermission(Manifest.permission.READ_EXTERNAL_STORAGE) != PackageManager.PERMISSION_GRANTED) {
                 ActivityCompat.requestPermissions(this, new String[]{Manifest.permission.READ_EXTERNAL_STORAGE}, REQUEST_CODE_PICK_IMAGE);
             }
         }
+ */
 
         Intent i = getIntent();
         pidInc= i.getIntExtra(TAG_IDINC, 0);
-        ptienda = i.getStringExtra(TAG_TIENDA);
-        /*
-        pidRuta = i.getIntExtra(TAG_IDRUTA, 0);
-        pidPromotor = i.getIntExtra(TAG_IDPROMOTOR, 0);
-        pLatitud = i.getDoubleExtra(TAG_LATITUD, 0.0);
-        pLongitud = i.getDoubleExtra(TAG_LONGITUD, 0.0);
-
-        pdireccion = i.getStringExtra(TAG_DIRECCION);
-        pfechahora = i.getStringExtra(TAG_FECHAHORA);
-        pfechahora_respuesta = i.getStringExtra(TAG_FECHAHORARESPUESTA);
-        prespuesta = i.getStringExtra(TAG_RESPUESTA);
-        pobservaciones= i.getStringExtra(TAG_OBSERVACIONES);
-         */
+        pfuente= i.getStringExtra(TAG_FUENTE);
+        Log.e(TAG_ERROR, "TAG_IDINC " + pidInc);
+        Log.e(TAG_ERROR, "TAG_FUENTE " + pfuente);
 
         oRespuestaIncidencia orespinc = almacenaImagen.obtenRespuestaIncidencia(pidInc);
-        // idincidencia, idfoto, idpromotor, idruta, fechahora, observaciones, respuesta, fechahora_respuesta, image, leida
+/*
         Log.e(TAG_ERROR, String.valueOf(pidInc));
-        Log.e(TAG_ERROR, String.valueOf(orespinc._idincidencia));
-        Log.e(TAG_ERROR, String.valueOf(orespinc._idfoto));
-        Log.e(TAG_ERROR, String.valueOf(orespinc._idruta));
-        Log.e(TAG_ERROR, orespinc._fechahora);
-        Log.e(TAG_ERROR, orespinc._observaciones);
-        Log.e(TAG_ERROR, orespinc._respuesta);
-        Log.e(TAG_ERROR, orespinc._fechahora_respuesta);
-        Log.e(TAG_ERROR, orespinc._image);
-        Log.e(TAG_ERROR, String.valueOf(orespinc._leida));
+        Log.e(TAG_ERROR, String.valueOf(oRespuestaIncidencia._idincidencia));
+        Log.e(TAG_ERROR, String.valueOf(oRespuestaIncidencia._idfoto));
+        Log.e(TAG_ERROR, String.valueOf(oRespuestaIncidencia._idruta));
+        Log.e(TAG_ERROR, String.valueOf(oRespuestaIncidencia._tienda));
+        Log.e(TAG_ERROR, String.valueOf(oRespuestaIncidencia._direccioncompleta));
+        Log.e(TAG_ERROR, oRespuestaIncidencia._fechahora);
+        Log.e(TAG_ERROR, oRespuestaIncidencia._observaciones);
+        Log.e(TAG_ERROR, oRespuestaIncidencia._respuesta);
+        Log.e(TAG_ERROR, oRespuestaIncidencia._fechahora_respuesta);
+        Log.e(TAG_ERROR, oRespuestaIncidencia._image);
+        Log.e(TAG_ERROR, String.valueOf(oRespuestaIncidencia._leida));
 
+ */
 
         //****************************
         // Agregando losdatos
+        TextView lblNotif = findViewById(R.id.lblNotif);
+        String textonotif = "Incidencia " + pidInc;
+        lblNotif.setText(textonotif);
+
+        //****************************
         TextView txtTituloTienda = findViewById(R.id.txtTienda);
-        txtTituloTienda.setText(ptienda.toUpperCase());
+        txtTituloTienda.setText(oRespuestaIncidencia._tienda.toUpperCase());
 
         // ***************************************
         // Obtiene el nombre del usuario en y promotor las preferencias
@@ -141,39 +117,26 @@ public class RespuestaIncidencia extends AppCompatActivity {
         TextView txtUsuario = findViewById(R.id.txtUsuario);
         txtUsuario.setText(pUsuario);
 
+        // *****************************
         TextView txtFechaHora = findViewById(R.id.txtFecha);
-        txtFechaHora.setText(orespinc._fechahora);
-
-        TextView txtfechahora_respuesta = findViewById(R.id.txtFechaRespuesta);
-        txtfechahora_respuesta.setText(orespinc._fechahora_respuesta);
-
-        TextView txtRespuesta = findViewById(R.id.txtRespuesta);
-        txtRespuesta.setText(orespinc._respuesta);
-
-        EditText txtObservaciones = findViewById(R.id.txtObservaciones);
-        txtObservaciones.setText(orespinc._observaciones);
+        txtFechaHora.setText(oRespuestaIncidencia._fechahora);
 
         // *****************************
-        // llenando los datos de la lista de productos
-        Cursor c = almacenaImagen.CursorIncidencias();
-        String[] from = new String[]{TAG_DESCRIPCIONINCIDENCIA};
-        int[] to = new int[]{android.R.id.text1};
-        SimpleCursorAdapter adapter = new SimpleCursorAdapter(this, android.R.layout.simple_spinner_item, c, from, to);
-        adapter.setDropDownViewResource(android.R.layout.simple_spinner_dropdown_item);
-        Spinner spinIncidencia = findViewById(R.id.spinIncidencia);
-        spinIncidencia.setAdapter(adapter);
-        spinIncidencia.setSelection(orespinc._idincidencia);
+        TextView txtfechahora_respuesta = findViewById(R.id.txtFechaRespuesta);
+        txtfechahora_respuesta.setText(oRespuestaIncidencia._fechahora_respuesta);
 
-        // ******************************
-        spinIncidencia.setOnItemSelectedListener(new AdapterView.OnItemSelectedListener() {
-            public void onItemSelected(AdapterView<?> parent, View view, int position, long id) {
-                idincidencia = Integer.parseInt(String.valueOf(id));
-                // Toast.makeText(getApplicationContext(),"Incidencia " + idincidencia, Toast.LENGTH_SHORT).show();
-            }
+        // *****************************
+        TextView txtRespuesta = findViewById(R.id.txtRespuesta);
+        txtRespuesta.setText(oRespuestaIncidencia._respuesta);
 
-            public void onNothingSelected(AdapterView<?> parent) {
-            }
-        });
+        // *****************************
+        EditText txtObservaciones = findViewById(R.id.txtObservaciones);
+        txtObservaciones.setText(oRespuestaIncidencia._observaciones);
+
+        // *****************************
+        // llenando de tipo de incidencia
+        TextView txtTipoIncidencia = findViewById(R.id.txtTipoIncidencia);
+        txtTipoIncidencia.setText(almacenaImagen.obtenTipoIncidencia(oRespuestaIncidencia._idincidencia));
 
         //****************************
         FloatingActionButton fab = findViewById(R.id.fabMenu);
@@ -192,18 +155,16 @@ public class RespuestaIncidencia extends AppCompatActivity {
         // Botón para guardar
         Button cmdCerrar = findViewById(R.id.cmdCerrar);
         // Boton foto puede tomar una foto
-        cmdCerrar.setOnClickListener(new View.OnClickListener() {
-            public void onClick(View view) {
-                almacenaImagen.estableceLeidorespuesta(pidInc);
-                Intent MenuTienda = new Intent(getApplicationContext(), listarespuestaincidencias.class);
-                MenuTienda.putExtra(TAG_IDRUTA, Integer.valueOf(pidRuta));
-                MenuTienda.putExtra(TAG_IDPROMOTOR, Integer.valueOf(pidPromotor));
-                MenuTienda.putExtra(TAG_TIENDA, ptienda);
-                MenuTienda.putExtra(TAG_LATITUD, pLatitud);
-                MenuTienda.putExtra(TAG_LONGITUD, pLongitud);
-                MenuTienda.putExtra(TAG_DIRECCION, pdireccion);
-                startActivity(MenuTienda);
-            }
+        cmdCerrar.setOnClickListener(view -> {
+            almacenaImagen.estableceLeidorespuesta(pidInc);
+            Intent MenuTienda = new Intent(getApplicationContext(), listarespuestaincidencias.class);
+            MenuTienda.putExtra(TAG_IDRUTA, Integer.valueOf(pidRuta));
+            MenuTienda.putExtra(TAG_IDPROMOTOR, Integer.valueOf(pidPromotor));
+            MenuTienda.putExtra(TAG_TIENDA, ptienda);
+            MenuTienda.putExtra(TAG_LATITUD, pLatitud);
+            MenuTienda.putExtra(TAG_LONGITUD, pLongitud);
+            MenuTienda.putExtra(TAG_DIRECCION, pdireccion);
+            startActivity(MenuTienda);
         });
 
         // *******************
@@ -220,15 +181,12 @@ public class RespuestaIncidencia extends AppCompatActivity {
             return;
         }
         fusedLocationClient.getLastLocation()
-                .addOnSuccessListener(this, new OnSuccessListener<Location>() {
-                    @Override
-                    public void onSuccess(Location location) {
-                        if (location != null) {
-                            // Use the location object
-                            pLatitud = location.getLatitude();
-                            pLongitud = location.getLongitude();
-                            // Do something with the location data
-                        }
+                .addOnSuccessListener(this, location -> {
+                    if (location != null) {
+                        // Use the location object
+                        pLatitud = location.getLatitude();
+                        pLongitud = location.getLongitude();
+                        // Do something with the location data
                     }
                 });
     }
